Change types of SNS checking
There are now three types of SNS check: normal (-S), strict (-St), and
unstrict (-Su).

Strict requires that all parts of the SNS are included in the BREX.

Normal will allow SNS codes of "0" to be omitted from the BREX when it
is the only code at that level.

Unstrict will allow any code to be used for an omitted level.