Fix usage of brexDmRef
The BREX referenced by the first data module was being used as the only
BREX for all subsequent data modules. Fixed so that each data module may
reference a different BREX.

If all data modules do reference the same BREX, use the -b option if
possible as it will be more efficient (the BREX or set of layered BREX
are only read once).