Add default dmtypes and samples
dmtypes and dmtypes.xml provide "default" samples using the standard
info codes and info names from Chap 8 of the S1000D specification.

dmtypes.xml is also built-in to the s1kd-newdm tool, meaning if no
project-specific dmtypes file is found or specified then it will be used
to provide a default schema and info name for the new data module.