Remove -t, -i, -l, -D opts, add ICNs and issueType
These options were removed in favour of using the dmlType attribute to
determine if a DML is a DMRL (so does not include issue info/date) or is
a CSL (includes issue info/date). Title and language are always
included.

The attribute issueType, derived from dmStatus, is added to dmlEntry
when the list is a CSL.

ICNs can be specified as arguments to be included in the list.