Use XSLT for CIR resolve, add -x option
In addition to allowing custom XSLT scripts for resolving CIR
dependencies (-r), a set of built-in XSLT scripts is included, which can
be dumped with the -x option as a starting point for custom scripts.

These also replace the C functions previously used to resolve
dependencies, and cover more types of CIRs.