Change behavior of arguments to s1kd-instance tool
Applicability definitions are now assigned using the -s option rather
than being included as the remaining arguments to the tool.

Source objects are now specified as the remaining arguments to the tool.
This allows multiple source objects to be specified when calling the
tool without needing to pass in a list (-L).