Search default BREX, add older issue support
The s1kd-brexcheck tool now includes the default S1000D BREX built-in.
These are only used as a "last resort" if the tool cannot locate a
copy of the default BREX in the current directory, any of the search
directories, or the input list of data modules.

Support is also added for older issues, 3.0, 2.3, and 2.2. Support for
older issues than those is not planned as the BREX had not been
introduced yet, although generally a BREX written for the supported
older issues should still work if specified manually.