s1kd-defaults: Prevent overwrite on re-initialize
The files created during initialization (-i) will not be overwritten if
the -i option is specified again, unless the -f option is specified.

The -b option will also cause the .defaults and .dmtypes files to be
overwritten without the -f option, but not the .fmtypes file. This allows
these files to be periodically updated from a BREX without discarding
changes to the .fmtypes file.