s1kd-defaults: Prevent auto overwrite with -ib
Do not overwrite on re-initialization, even if the -b option is
specified. There is information loss between BREX <-> defaults files, so
care must always be taken when updating from the BREX.

The -f option is required to overwrite these files, otherwise it will
only recreate those that do not exist.