Separate doc build/clean from binaries
Because the docs are generated ahead of time and with non-standard
tools, normal users shouldn't have to build them.

Before, we relied on make to not rebuild the docs if they already
existed after cloning the repo, but make uses the timestamp of a file to
determine if the recipe needs to be run again, and the cloning process
gives somewhat random results in terms of file timestamps, depending on
the order files are written to disk. And users could not do a 'clean'
without deleting the pre-built docs.

The better way is to have separate recipes for building/cleaning the
binaries vs. the docs:

- 'make' or 'make all' will now only build binaries.
- 'make doc' will build all the documentation.
- 'make clean' will only clean binaries
- 'make maintainer-clean' will clean docs and binaries