Change behaviour of s1kd-instance -a option
The -a option will now remove unused applicability statements as well as
the annotations they correspond to. This is only if the entire statement
is invalid to the given applicability. It remains distinguished from the
-A option in that the latter will also attempt to simplify statements
by removing parts of them that are invalid.

The documentation is also updated with a more detailed example of the
differences between these options.