Show errors on stderr and sooner
The BREX/SNS/NOTATION ERROR messages will now be printed to stderr
instead of stdout. They will also be shown even when -f or -x are given,
since they will not interfere with other stdout output. To disable them,
use the -q option.

The errors and filenames (-f) will now be printed as soon as they are
encountered during the check, instead of only after it completes.