Add errors and warnings for missing BREX DMs
If a CSDB object references a BREX DM, but that DM is not found, the check
will immediately fail.

If a CSDB object does not reference a BREX DM (optional in some schemas,
especially in older issues), and no BREX DMs are explicitly specified
to check objects against (-b or -B), the tool will emit a warning,
the object will be ignored and the check will continue.