Change s1kd-appcheck -p to -t, add new -p option
Changed s1kd-appcheck to work more like s1kd-brexcheck. Specified
objects are read in to a list in memory before checking. One advantage
of this is that referenced DMs (ACT, CCT, PCT) can be found among the
specified objects if they cannot be found in the directory hierarchy.

Because of this, there can also be a -p option like s1kd-brexcheck to
display a progress bar. The current -p option was renamed to -t.