Fix IDSTATUS handling when updating instances
Originally, the IDSTATUS section of an instance was copied when
updating it from the master, as a simple means of keeping the same
metadata. However, it is possible for information in the IDSTATUS section
to be subject to different applicability, for example, reasons for update
and QA status. Therefore, the IDSTATUS must be updated from  the master
as well.

Now, the IDSTATUS section is no longer copied. Instead, the metadata
needed to produce the instance is fetched and re-applied to the master
when updating.