s1kd-instance: Do not mix asserts with -@/-8
The -@/-8 options will add the applicability of an instance/source to
the user-defined filters. Previously, if these contained assertions
that conflicted with the user-defined assertions, the assertions would
be combined. However, this is not the desired behaviour, and can cause
issues when -8 and -w are combined. Now, the user-defined assertions
will override the assertions in the instance/source.