s1kd-instance: Fix -y with -8
Part of the -y option implementation simplifies the applic of the
whole object before combining it with the user-definitions, to remove
redundant definitions. However, the simplifying process did not ignore
the definitions introduced by the -8 option, and would therefore always
result in the applic of the whole object being simplified to "All". This
has been fixed so that only the user-defined applicability is considered
when simpliying the whole object's applicability.