s1kd-instance: Set default issue type with -n
When the -n option is used to set the issue of an instance, the instance
will also be assigned an appropriate default issue type.

If the issue number is 000-01 thru 001-00, the default issue type will
be "new" (this behaviour was already implemented).

If the issue number is 001-01 and up, the default issue type will be
"changed" (this behaviour is new).

The default issue type can still be overridden with the -z option.