s1kd-upissue: Add -0 (--unverified) option
Added a -0 (--unverified) option to set an object's QA status to
unverified.

Changed the behaviour of the -q option to remove its alternative
function in -m mode. This is replaced by -0. Consequently, the secondary
long option (--reset-qa) is also removed.

Changed the behaviour of the -1 option to unset the 2nd verification
type of objects when the -2 option is not specified.

Changed the behaviour of the -2 option to automatically set a 1st
verification type if the object is unverified and -1 is not specified.
The type will be the same as that specified with -2.