s1kd-brexcheck: Link with libsaxonhec
Modified the Saxon/C API to link with libsaxonhec at compile-time
instead of loading it dynamically at run-time, in order to work around
issues with the JET runtime on Windows/Cygwin.

Only the bare-minimum of the dynamic loading code was commented out, so
there is likely plenty more unused code that can be cleaned up.