s1kd-brexcheck: Add dynamic XPath support
If an XPath 2.0 engine is selected at compile-time, s1kd-brexcheck will
now dynamically use the appropriate version of XPath based on the S1000D
issue of the BREX data module (3.0 and lower = XPath 1.0, 4.0 and up =
XPath 2.0).

The -X (--xpath-version) option allows the user to force a particular
version of XPath, regardless of the issue.

The Makefile variable BREXCHECK_XPATH_ENGINE was renamed to
XPATH2_ENGINE accordingly.