s1kd-appcheck: Add -#, --threads option
Replaced the -j (--jobs) option with -# (--threads), with the intention of
eventually adding this same option to other tools such as s1kd-validate,
s1kd-brexcheck and s1kd-instance.

Also expanded the syntax of the argument to <x[,y]>, where x represents
the number of threads created to iterate over the list of objects,
and y represents the number of threads created to validate each object
(meaning x * y threads are created in total).