s1kd-flatten: Remove "merging" behaviour of -a/-R
Removed the "merging" behaviour of the -R and recently added -a options,
where the contents of the referenced PM would be copied into the main
PM in an S1000D-compliant way. Instead, the tool will always use the
non-S1000D-compliant format of including the whole <pm> element in place
of the <pmRef>.

There are too many different ways that merging can be done, so it's
better to let projects decide how using a separate tool, such as an
XSL transformation on the output of s1kd-flatten. For example, they can
decide to make the output S1000D-compliant by turning the <pm> element
into a <pmEntry>, or copying the contents directly into the main PM.

The -a option has been removed since it no longer has any meaning with
the new behaviour of -R.