s1kd-validate: Shallow copy nodes to XML report
Changed the way nodes are copied to the XML report so that it will only
be a shallow copy of the node (only attributes, no children). This mimics
the s1kd-brexcheck XML report, and prevents the report from becoming
very large if a node with a lot of content is reported.