s1kd-instance: Fix bug when loading applic from instance/object
Fixed a bug where loading the applicability definitions from an instance
or other object may result in incorrect logic being applied.

The applicability definitions mechanism currently does not support the
"or" operator, so assertions in "or" evaluations must be ignored when
loading the applicability definitions from the object.

This affects the -H (--list-properties), -8 (--reapply), and -@
(--update-instances) options.