s1kd-instance: Improve behaviour of -H applic
Improved the behaviour of the -H applic option, which lists the
applicability properties and specific values that are applicable to a
given data module.

Instead of using the load_applic_from_inst function, which is mainly used
for the -@ (--update-instances) option but is more limited since the
internal applic defs construct cannot fully handle "or" statements, it
now checks each value against the data module applicability to determine
whether to add it to the list of applicable values.

This should allow the list of values to be tailored more accurately even
for more complex applicability.