Ignore deprecated warning for xmlAddNotationDecl
In libxml >= 2.14.0, the function xmlAddNotationDecl has been marked as
deprecated/internal use only. However, I can't find another function to
add a NOTATION to the DTD.

See: https://discourse.gnome.org/t/how-to-add-a-notation-to-a-dtd/30557

If another function is found/added later, it should be used instead.

Also bumped the versions of all tools that use this function via
add_notation or add_icn.