Remove unneeded xmlValidCtxt and update comment
Removed the unneeded creation of an xmlValidCtxt for the use of the
xmlAddNotationDecl function, using NULL instead.

Updated the comment regarding the pragmas around xmlAddNotationDecl to
include a link to the issue raised on libxml's repository. The version of
libxml where the function has been undeprecated will be determined later.