Remove deprecated use of xmlNoNetEntityLoader
libxml prior to version 2.15.0 included a built-in HTTP client that
would be used to load entities, so it was necessary to disable this when
not using the --net option to allow network connections by setting the
default entity loader function to xmlNoNetEntityLoader.

As of version 2.15.0, the built-in HTTP client has been removed from
libxml, so setting the loader function is no longer necessary to disable
network connections.